+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Redux Framework Changelog
 
+## 3.3.9.17
+* Added     SASS compiler for admin CSS.
+* Added     SASS to CSS for grunt compiler.
+
 ## 3.3.9.16
 * Added     SASS compiler for fields.